feat(todo): redesigned ToDo tab with categorized habit composer
- New ToDoView with editorial typography header (no name, two-line greeting)
- Composer redesigned: 5 habit categories (Move/Mind/Care/Connect/Home), library tiles, custom input
- Removed time/date noise from task cards and detail view
- Bumped all rounded surfaces to 28pt for visual consistency
- Past Letters tile typography updated, horizontal scroll
- Reading hero card and Me tab Reading section redesign
- Bond cards in People tab redesigned to be visually distinct
- Swift 5.0 + minimal concurrency to fix AVFoundation strict-concurrency errors

+++ Packages/Core/DesignSystem/Components/NyraOrb.swift
@@ -933,8 +933,8 @@ private struct BlobShape: Shape {
             let radial = 1.0 + CGFloat(wobble1 + wobble2 * 0.6) * bulge * 0.5
             let rr = r * radial
             points.append(CGPoint(
-                x: cx + cos(theta) * rr,
-                y: cy + sin(theta) * rr
+                x: cx + CGFloat(cos(theta)) * rr,
+                y: cy + CGFloat(sin(theta)) * rr
             ))
         }
